+++ src/api/index.ts
@@ -1,9 +1,5 @@
-export {
-  getTodos,
-  getTodoById,
-  createTodo,
-  updateTodo,
-  deleteTodo,
-} from './todos'
+export { getTodos, getTodoById, deleteTodo } from './todos'
 
 export { login, signUp } from './user'
+
+export { api } from './api'
