+++ src/components/TodoEdit/api/createTodo.ts
@@ -1,0 +1,6 @@
+import { TodoInput, TodoResponse } from 'types'
+import { api } from 'api'
+
+type CreateTodoParam = TodoInput
+export const createTodo = (json: CreateTodoParam) =>
+  api().post('todos', { json }).json<TodoResponse>()
