+++ src/components/todo/index.ts
@@ -1,4 +1,3 @@
 export * from './TodoItem'
 export * from './TodoPreview'
 export * from './TodoSkeleton'
-export * from './AddTodo'
