+++ src/queries/todo.ts
@@ -1,5 +1,5 @@
 import { UseQueryOptions, useMutation, useQuery } from '@tanstack/react-query'
-import { createTodo, deleteTodo, getTodos, updateTodo } from 'api'
+import { deleteTodo, getTodos } from 'api'
 import { Todo } from 'types'
 import { queryClient } from 'queryClient'
 
@@ -20,16 +20,6 @@ export const useTodoQuery = (id: string) =>
   useQuery({
     ...todosQueryOption,
     select: data => data.find(todo => todo.id === id),
-  })
-
-export const useCreateTodoMutation = () =>
-  useMutation(createTodo, {
-    onSuccess: () => queryClient.invalidateQueries(['todos']),
-  })
-
-export const useUpdateTodoMutation = () =>
-  useMutation(updateTodo, {
-    onSuccess: () => queryClient.invalidateQueries(['todos']),
   })
 
 export const useDeleteTodoMutation = () =>
